+++ src/database/initialization/changemaker_field_value_batch_to_json.sql
@@ -1,13 +1,19 @@
 SELECT drop_function('changemaker_field_value_batch_to_json');
 
 CREATE FUNCTION changemaker_field_value_batch_to_json(
-	changemaker_field_value_batch changemaker_field_value_batches
+	changemaker_field_value_batch changemaker_field_value_batches,
+	auth_context_keycloak_user_id uuid DEFAULT NULL,
+	auth_context_is_administrator boolean DEFAULT FALSE
 )
 RETURNS jsonb AS $$
 DECLARE
 	source_json JSONB;
 BEGIN
-	SELECT source_to_json(sources.*)
+	SELECT source_to_json(
+		sources.*,
+		auth_context_keycloak_user_id,
+		auth_context_is_administrator
+	)
 	INTO source_json
 	FROM sources
 	WHERE sources.id = changemaker_field_value_batch.source_id;