+++ src/database/queries/changemakerFieldValueBatches/selectById.sql
@@ -1,6 +1,8 @@
 SELECT
 	changemaker_field_value_batch_to_json(
-		changemaker_field_value_batches.*
+		changemaker_field_value_batches.*,
+		:authContextKeycloakUserId,
+		:authContextIsAdministrator
 	) AS object
 FROM changemaker_field_value_batches
 WHERE