+++ src/database/queries/proposalVersions/insertOne.sql
@@ -18,4 +18,8 @@ INSERT INTO proposal_versions (
 		1
 	)
 )
-RETURNING proposal_version_to_json(proposal_versions) AS object;
+RETURNING proposal_version_to_json(
+	proposal_versions,
+	:authContextKeycloakUserId,
+	:authContextIsAdministrator
+) AS object;