+++ src/database/queries/sources/deleteOne.sql
@@ -1,4 +1,8 @@
 DELETE FROM sources
 WHERE
 	id = :sourceId
-RETURNING source_to_json(sources) AS object;
+RETURNING source_to_json(
+	sources,
+	:authContextKeycloakUserId,
+	:authContextIsAdministrator
+) AS object;