+++ src/database/queries/sources/insertOne.sql
@@ -10,4 +10,8 @@ VALUES (
 	:funderShortCode,
 	:changemakerId
 )
-RETURNING source_to_json(sources) AS object;
+RETURNING source_to_json(
+	sources,
+	:authContextKeycloakUserId,
+	:authContextIsAdministrator
+) AS object;