+++ src/database/queries/sources/selectById.sql
@@ -1,3 +1,8 @@
-SELECT source_to_json(sources.*) AS object
+SELECT
+	source_to_json(
+		sources.*,
+		:authContextKeycloakUserId,
+		:authContextIsAdministrator
+	) AS object
 FROM sources
 WHERE id = :sourceId;